LPS-169174 Move logic to mgmt toolbar display context
